MCP-Automation-Sandbox/blaxel-swagger-finder/requirements.txt
blaxel
requests
streamlit
watchdog
httpx
pyyaml
pydantic
python-dotenv
fastapi
uvicorn
